Split domain code by feature
