+++ src/main/scala/io/github/pauljamescleary/petstore/domain/users/UserValidationInterpreter.scala
@@ -1,11 +1,9 @@
-package io.github.pauljamescleary.petstore.domain.validation.interpreters
+package io.github.pauljamescleary.petstore.domain.users
 
 import cats._
 import cats.data.EitherT
 import cats.implicits._
-import io.github.pauljamescleary.petstore.domain.model.User
-import io.github.pauljamescleary.petstore.domain.repository.UserRepositoryAlgebra
-import io.github.pauljamescleary.petstore.domain.validation.{UserAlreadyExistsError, UserValidationAlgebra}
+import io.github.pauljamescleary.petstore.domain.UserAlreadyExistsError
 
 class UserValidationInterpreter[F[_]: Monad](userRepo: UserRepositoryAlgebra[F]) extends UserValidationAlgebra[F] {
   def doesNotExist(user: User) = EitherT {
