+++ src/main/scala/io/github/pauljamescleary/petstore/infrastructure/repository/inmemory/PetRepositoryInMemoryInterpreter.scala
@@ -6,8 +6,7 @@ import scala.util.Random
 import cats._
 import cats.data.NonEmptyList
 import cats.implicits._
-import io.github.pauljamescleary.petstore.domain.model.{Pet, PetStatus}
-import io.github.pauljamescleary.petstore.domain.repository.PetRepositoryAlgebra
+import io.github.pauljamescleary.petstore.domain.pets.{Pet, PetRepositoryAlgebra, PetStatus}
 
 class PetRepositoryInMemoryInterpreter[F[_]: Applicative] extends PetRepositoryAlgebra[F] {
